backend/requirements.txt
fastapi
uvicorn
websockets
azure-cognitiveservices-speech
python-dotenv
